adding listing details unit tests for service and controller

+++ src/main/java/com/skillstorm/lit/services/ListingDetailsServiceImpl.java
@@ -4,8 +4,12 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 
+import org.apache.logging.log4j.LogBuilder;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +19,7 @@ import com.skillstorm.lit.models.ListingDetails;
 @Service
 @Transactional
 public class ListingDetailsServiceImpl implements ListingDetailsService {
+	private final Logger LOG = LoggerFactory.getLogger(getClass());
 	
 	@Autowired
 	ListingDetailsRepository repository;
@@ -34,13 +39,25 @@ public class ListingDetailsServiceImpl implements ListingDetailsService {
 
 	@Override
 	public ListingDetails update(ListingDetails detail, UUID id) {
+		LOG.debug("===================================================\n" + 
+	              "===================================================\n" +
+				  "updating ListingDetails entity with id = " + id +
+				  "===================================================");
 		detail.setId(id);
-		return repository.save(detail);
+		if (repository.findById(id).isPresent()) {
+			return repository.save(detail);
+		} else {
+			throw new EntityNotFoundException();
+		}
 	}
 
 	@Override
 	public void delete(UUID id) {
-		repository.deleteById(id);
+		if (repository.findById(id).isPresent()) {
+			repository.deleteById(id);
+		} else {
+			throw new EntityNotFoundException();
+		}
 	}
 
 }
